# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Screening Room API — TMDB proxy and notifications
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MediaCardMediaType = typeof MediaCardMediaType[keyof typeof MediaCardMediaType];


export const MediaCardMediaType = {
  movie: 'movie',
  tv: 'tv',
} as const;

export interface MediaCard {
  tmdb_id: number;
  media_type: MediaCardMediaType;
  title: string;
  /** @nullable */
  year?: number | null;
  genres: string[];
  /** @nullable */
  poster_path?: string | null;
  /** @nullable */
  vote_average?: number | null;
  /** @nullable */
  popularity?: number | null;
}

export interface CastMember {
  id: number;
  name: string;
  character: string;
  /** @nullable */
  profile_path?: string | null;
  order: number;
}

export interface VideoResult {
  key: string;
  name: string;
  site: string;
  type: string;
}

export type MediaDetailMediaType = typeof MediaDetailMediaType[keyof typeof MediaDetailMediaType];


export const MediaDetailMediaType = {
  movie: 'movie',
  tv: 'tv',
} as const;

/**
 * @nullable
 */
export type MediaDetailNextEpisodeToAir = { [key: string]: unknown } | null;

export interface MediaDetail {
  tmdb_id: number;
  media_type: MediaDetailMediaType;
  title: string;
  /** @nullable */
  overview?: string | null;
  /** @nullable */
  year?: number | null;
  genres: string[];
  /** @nullable */
  poster_path?: string | null;
  /** @nullable */
  backdrop_path?: string | null;
  /** @nullable */
  vote_average?: number | null;
  /** @nullable */
  runtime?: number | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  tagline?: string | null;
  cast: CastMember[];
  trailer?: VideoResult | null;
  /** @nullable */
  number_of_seasons?: number | null;
  /** @nullable */
  next_episode_to_air?: MediaDetailNextEpisodeToAir;
}

export interface EpisodeCheckInput {
  tmdb_ids: number[];
}

export interface EpisodeAlert {
  tmdb_id: number;
  title: string;
  season_number: number;
  episode_number: number;
  episode_name: string;
  air_date: string;
  /** @nullable */
  poster_path?: string | null;
}

export type SearchMediaParams = {
query: string;
type?: SearchMediaType;
};

export type SearchMediaType = typeof SearchMediaType[keyof typeof SearchMediaType];


export const SearchMediaType = {
  all: 'all',
  movie: 'movie',
  tv: 'tv',
} as const;