# lib/api-zod/src/generated/types/castMember.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Screening Room API — TMDB proxy and notifications
 * OpenAPI spec version: 0.1.0
 */

export interface CastMember {
  id: number;
  name: string;
  character: string;
  /** @nullable */
  profile_path?: string | null;
  order: number;
}