# lib/api-zod/src/generated/types/mediaDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Screening Room API — TMDB proxy and notifications
 * OpenAPI spec version: 0.1.0
 */
import type { CastMember } from './castMember';
import type { MediaDetailMediaType } from './mediaDetailMediaType';
import type { MediaDetailNextEpisodeToAir } from './mediaDetailNextEpisodeToAir';
import type { VideoResult } from './videoResult';

export interface MediaDetail {
  tmdb_id: number;
  media_type: MediaDetailMediaType;
  title: string;
  /** @nullable */
  overview?: string | null;
  /** @nullable */
  year?: number | null;
  genres: string[];
  /** @nullable */
  poster_path?: string | null;
  /** @nullable */
  backdrop_path?: string | null;
  /** @nullable */
  vote_average?: number | null;
  /** @nullable */
  runtime?: number | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  tagline?: string | null;
  cast: CastMember[];
  trailer?: VideoResult | null;
  /** @nullable */
  number_of_seasons?: number | null;
  /** @nullable */
  next_episode_to_air?: MediaDetailNextEpisodeToAir;
}